Add method add_turn
I changed the way turns are handled a bit - the turn number
is the number of the current turn, which is zero after the first turn
has been added.

Previously, the "current" turn didn't exist in the database, but I
realized this would lead to problems. Now it is created before anything
needs to be done with it.